Fix imports - Incorrect multiple redefinition
Each sub-package was redefining each type for cgo causing them to
not operate with each other.

ex. a CodecContext created in the avformat package was not actually the
same (and therefore could not use) the methods in the avcodec/context_struct.go file.

This change includes removing a lot of the incorrect redefinitions.
Also added another basic example thats correct and shows a bit  more.